+++ packages/ratel/lib/core/parameter_location.dart
@@ -1,9 +1,10 @@
 enum ParameterLocation {
   path,
-
   query,
-
   header,
-
   cookie,
+  body,
+  context,
+  multipart,
+  webSocket,
 }