+++ packages/ratel/test/support/fixtures/controllers/greeter_controller.dart
@@ -1,0 +1,9 @@
+import 'package:ratel/ratel.dart';
+
+final class GreeterController {
+  GreeterController(this.salutation);
+
+  final String salutation;
+
+  Future<Response> greet() async => Response.text(data: salutation);
+}